# Reimplement algorithms using TypeSortedCollection, perf improvements