Reimplement the following algorithms using `TypeSortedCollection`:
* `mass_matrix[!]`
* `geometric_jacobian[!]`
* `momentum_matrix[!]`
* `constraint_jacobian_and_bias!`

Addresses the main issue with https://github.com/tkoolen/RigidBodyDynamics.jl/pull/268 (reuse of motion subspaces transformed to world frame). Still changes the result of `geometric_jacobian` from the 'compact' representation (acting only on `velocity(state, path)`) to the 'full' representation (acting on `velocity(state)`), because I've found this easier to work with in practice (for setting up QPs and such) and because it is easier to implement in combination with `TypeSortedCollection`. Adds an `IntSet` field to `TreePath` to make membership tests in `geometric_jacobian!` efficient.

Changes the return type of `motion_subspace(joint, q)` and `constraint_wrench_subspace(joint, transform)` to  `GeometricJacobian{SMatrix{3, N, ...}}` and `WrenchMatrix{SMatrix{3, N, ...}}` respectively, where `N` is the number of degrees of freedom of the joint's `JointType`. The return types of these functions used to be `MotionSubspace` and `WrenchSubspace`, type aliases for 'fixed-maximum-size' versions of `GeometricJacobian` and `WrenchMatrix`, so as to make `motion_subspace` and `constraint_wrench_subspace` type stable in conjunction with the old [`@rtti_dispatch`](https://github.com/tkoolen/RigidBodyDynamics.jl/commit/657864eeef0e8dc2a735bc510574bf2a02533154#diff-e59f46cfbe787514a9a0681f29a05076L88) macro. I'm still storing the transformed motion subspaces and constraint wrench subspaces as `MotionSubspace`s and `WrenchSubspace`s in `MechanismState` for now, but any computations involving these data now act on fixed-size matrices, as opposed to using my half-baked type-pirated fixed-maximum-size matrix library, which is now gone.

Adds a benchmark for `constraint_jacobian_and_bias!`.

Also do some less invasive performance improvements (mostly eliding unnecessary cache checks).

Use StaticArrays master on Travis for now because of https://github.com/JuliaArrays/StaticArrays.jl/pull/251.

| Algorithm        | Before (μs)           | After (μs)  |
| ------------- |-------------:| -----:|
| `dynamics!` |  56.151 | 46.064 |
|  `momentum_rate_bias` | 12.401 | 9.316 |
|   `mass_matrix!` | 22.271 | 12.709 |
|   `momentum_matrix!` | 14.047 | 8.147 |
|   `geometric_jacobian!` | 10.701 | 5.172 |
|   `momentum` | 7.546 | 6.158 |
|   `kinetic_energy` | 7.168 | 5.688 |